requirements.txt
dj-database-url==0.5.0
Django==4.0.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
wheel==0.37.1
whitenoise==6.0.0
python-decouple==3.6
django-ckeditor==6.5.0
